move SyncState to domain
