+++ crates/domain/src/models/mod.rs
@@ -6,6 +6,7 @@ pub mod execution_payload_cache;
 pub mod peer_list;
 pub mod reth_provider;
 pub mod storage_module;
+pub mod sync_state;
 
 pub use block_index::*;
 pub use block_tree::*;
